third_party/WebKit/LayoutTests/fast/dom/Document/embeds-non-html.html

<!DOCTYPE html>
<html>
<head>
<link rel="help" href="http://www.whatwg.org/specs/web-apps/current-work/multipage/dom.html#dom-document-embeds"/>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>
description("Make sure the embeds / plugins attributes are exposed on non-HTML documents.");

var xmlDocument = document.implementation.createDocument("", "", null);
shouldBe("xmlDocument.__proto__", "XMLDocument.prototype");
shouldBe("xmlDocument.__proto__.__proto__", "Document.prototype");

shouldBe("xmlDocument.embeds.__proto__", "HTMLCollection.prototype");
shouldBe("xmlDocument.embeds.length", "0");
shouldBe("xmlDocument.embeds", "xmlDocument.plugins");
</script>
</body>
</html>
