third_party/WebKit/LayoutTests/fast/dom/Document/embeds-non-html-expected.txt

Make sure the embeds / plugins attributes are exposed on non-HTML documents.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS xmlDocument.__proto__ is XMLDocument.prototype
PASS xmlDocument.__proto__.__proto__ is Document.prototype
PASS xmlDocument.embeds.__proto__ is HTMLCollection.prototype
PASS xmlDocument.embeds.length is 0
PASS xmlDocument.embeds is xmlDocument.plugins
PASS successfullyParsed is true

TEST COMPLETE
